Started moving agent grid to new file created helpers.py
